refactor(apps/web/인게임): InGamePage로 UI 로직을 이전
